requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
tqdm>=4.65.0
colorama>=0.4.6
click>=8.1.0
